refactor: consolidate weapon handling into a single generic handler; remove individual shot handlers
